VM: Switch from Rc to Ref with Lifetime
As the individual Ops will need to modify the context, I need them
to be mutable. However Rc doesnt allow mutable reference to the
underlying data to be extracted. So change to maintaining the context
as a mutable reference along with a associated lifetime.

As we are not looking at multi threaded operations, and as instructions
will execute serially, this is the simpler flow.

TODO: Rather come to think of it, there is no need to store the context
within the Op, rather when running the Op, the context can be passed
to it. So may be will avoid embedding context within Op