VM:Op:IOBridgeRead resizes buf to read length
Also read will read only upto fill the buf passed to it and not
more, So after the 1st/any read resizes the buffer, subsequent
reads wont read more than that amount of data.

So also recreate the buf each time through the loop, so that we
can read how much ever we want to read, as specified with bufsize
to bufnew operation in the program.