Initial skeleton of DataManager enum
Currently it handles isize int data of the VM.

It allows a int data element to be either a int literal or a int
var.

NOTE: In compile I have left skeleton for a DataM to handle isize
for now or say string/u64/... in future. However In Run I have
forced it to return isize for now.