VM:Compile:Op:If??.s: Allow string literal with space
Till now one could either use a String Var or A StringValue in
double quotes but without spaces in it, wrt If conditions two
values being checked against one another.

However with this commit, one should be able to use any double
quoted string as part of the if condition check.

Also added a simple test.if.prg to test if condition checking.