VM:DataM:Compile: Fix issue with finding FuncArg data type
I was wrongly using the runtime fargs aliases hashmap instead of
the compile time funcs hashmap, when trying to check if the var
being DataM'd is funcarg or not. This is fixed now.

In order to achieve the same, a new compilingfunc (func name) data
member has been added and managed during compiling.