VM:Op:IOBRead: Switch to new DataM based flow wrt dest
NOTE: I am not directly updating the underlying buf, passed to
iobread. Rather I update a clone and then reset the variable to
the new updated buffer.

Even thou not the ideal flow. For now using it, bcas the helpers
setup wrt getting the underlying mutable buffer, borrows the ctxt
var in a mutable way. And before its use is finished, I also try
to directly use ctxt again as a mutable reference to get a mut
ref to the iobridge. And this doesnt compute in Rust by default,
even thou technically in this case it doesnt create a issue
theoretically.

Even if iobread where to allow reading into the middle of a buffer,
this should be fine. As we are duplicating the original buf,
rather than creating a new fresh buf.